Cast to long

+++ AppCenterDataStorage/AppCenterDataStorage/MSDataStore.m
@@ -392,7 +392,8 @@ static dispatch_once_t onceToken;
                                     NSError *actualError = [MSCosmosDb cosmosDbErrorWithResponse:response underlyingError:cosmosDbError];
                                     MSLogError([MSDataStore logTag],
                                                @"Unable to list documents for partition %@: %@. Status code %ld when expecting %ld.",
-                                               partition, [actualError localizedDescription], response.statusCode, MSHTTPCodesNo200OK);
+                                               partition, [actualError localizedDescription], (long)response.statusCode,
+                                               (long)MSHTTPCodesNo200OK);
                                     MSDataSourceError *dataSourceCosmosDbError = [[MSDataSourceError alloc] initWithError:actualError];
                                     MSPaginatedDocuments *documents = [[MSPaginatedDocuments alloc] initWithError:dataSourceCosmosDbError
                                                                                                         partition:partition
@@ -501,7 +502,7 @@ static dispatch_once_t onceToken;
                                 NSError *actualError = [MSCosmosDb cosmosDbErrorWithResponse:response underlyingError:cosmosDbError];
                                 MSLogError([MSDataStore logTag],
                                            @"Unable to read document %@ with error: %@. Status code %ld when expecting %ld.", documentId,
-                                           [actualError localizedDescription], response.statusCode, MSHTTPCodesNo200OK);
+                                           [actualError localizedDescription], (long)response.statusCode, (long)MSHTTPCodesNo200OK);
                                 completionHandler([[MSDocumentWrapper alloc] initWithError:actualError documentId:documentId]);
                               }
 
@@ -553,8 +554,8 @@ static dispatch_once_t onceToken;
                               NSError *actualError = [MSCosmosDb cosmosDbErrorWithResponse:response underlyingError:cosmosDbError];
                               MSLogError([MSDataStore logTag],
                                          @"Unable to create/replace document %@ with error: %@. Status code %ld when expecting %ld or %ld.",
-                                         documentId, [actualError localizedDescription], response.statusCode, MSHTTPCodesNo200OK,
-                                         MSHTTPCodesNo201Created);
+                                         documentId, [actualError localizedDescription], (long)response.statusCode,
+                                         (long)MSHTTPCodesNo200OK, (long)MSHTTPCodesNo201Created);
                               completionHandler([[MSDocumentWrapper alloc] initWithError:actualError documentId:documentId]);
                             }
 
@@ -584,7 +585,7 @@ static dispatch_once_t onceToken;
                                 NSError *actualError = [MSCosmosDb cosmosDbErrorWithResponse:response underlyingError:cosmosDbError];
                                 MSLogError([MSDataStore logTag],
                                            @"Unable to delete document %@ with error: %@. Status code %ld when expecting %ld.", documentId,
-                                           [actualError localizedDescription], response.statusCode, MSHTTPCodesNo204NoContent);
+                                           [actualError localizedDescription], (long)response.statusCode, (long)MSHTTPCodesNo204NoContent);
                                 completionHandler([[MSDocumentWrapper alloc] initWithError:actualError documentId:documentId]);
                               }
